+++ src/main/java/com/timora/app/controllers/DisponibilidadController.java
@@ -24,12 +24,12 @@ public class DisponibilidadController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Disponibilidad> getById(@PathVariable Integer id) {
+    public ResponseEntity<Disponibilidad> getById(@PathVariable Long id) {
         return ResponseEntity.ok(disponibilidadService.findById(id));
     }
 
     @GetMapping("/proveedor/{idProveedor}")
-    public ResponseEntity<List<Disponibilidad>> getByProveedor(@PathVariable Integer idProveedor) {
+    public ResponseEntity<List<Disponibilidad>> getByProveedor(@PathVariable Long idProveedor) {
         return ResponseEntity.ok(disponibilidadService.findByProveedor(idProveedor));
     }
 
@@ -40,13 +40,13 @@ public class DisponibilidadController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Disponibilidad> actualizar(@PathVariable Integer id,
+    public ResponseEntity<Disponibilidad> actualizar(@PathVariable Long id,
                                                      @RequestBody Disponibilidad disponibilidad) {
         return ResponseEntity.ok(disponibilidadService.actualizar(id, disponibilidad));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> borrar(@PathVariable Integer id) {
+    public ResponseEntity<Void> borrar(@PathVariable Long id) {
         disponibilidadService.borrar(id);
         return ResponseEntity.noContent().build();
     }
